feat!: add spaghetti, steak frites, apple pie recipes
Introduce three recipes: Spaghetti, Steak Frites, and Apple Pie to be
displayed on Recipe site, specifically 3 html files in the recipes
folder, 3 .jpg images in the images folder, use of lists, images, and
links, and experimental use of two ordered lists in the Steak Frites
recipe for the two parts of the meal.

The README file remains to be updated.